requirements.txt
Django==5.1.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
geopy==2.4.1
Pillow==10.3.0
python-dotenv==1.0.1
gunicorn==21.2.0
whitenoise==6.6.0